refactor: opt-in credential resolution via from_default(); rename mixin module
- bare EESession() now raises; resolution is opt-in via EESession.from_default()
  (deferred, so async callers stay non-blocking)
- rename sepal_credential_mixin -> credential_mixin (thin BC shim kept)
- EESession subclasses CredentialMixin directly; tidy create() docstring; docs
