+++ tests/test_agnostic_auth_integration.py
@@ -1,10 +1,9 @@
-"""Gated integration test: headerless EESession() from EARTHENGINE_TOKEN.
+"""Gated integration test: EESession.from_default() from EARTHENGINE_TOKEN.
 
 Named to avoid conftest's ``test_integration_*`` collect-ignore glob, and skipped
 unless EARTHENGINE_TOKEN is present (set in CI, typically a service account).
 """
 
-import asyncio
 import os
 
 import pytest
@@ -17,19 +16,20 @@ pytestmark = pytest.mark.skipif(
 )
 
 
-def test_headerless_session_authenticates_from_earthengine_token(monkeypatch, tmp_path):
+def test_from_default_authenticates_from_earthengine_token(monkeypatch, tmp_path):
     # Non-SEPAL context (non-sepal-user home). SEPAL_HOST may be set; it must NOT
     # gate the token path (D11). Isolate home so no local cred file interferes.
     monkeypatch.setattr("eeclient.providers.Path.home", lambda: tmp_path)
     (tmp_path / ".config/earthengine").mkdir(parents=True)
 
-    session = EESession()  # bare -> resolves EARTHENGINE_TOKEN, refresh deferred
-    assert session._credentials is None  # not refreshed at construction (D6)
+    session = EESession.from_default()  # explicit env resolution, refresh deferred
+    assert session._credentials is None  # non-blocking construction (deferred)
 
-    headers = asyncio.run(session.get_headers()).model_dump(by_alias=True)
+    session.set_credentials_sync()  # trigger the real token mint
+    assert session.access_token
+    assert session.project_id
+    headers = session.get_current_headers().model_dump(by_alias=True)
     assert headers["Authorization"].startswith("Bearer ")
-    assert headers["x-goog-user-project"]
-    asyncio.run(session.aclose())
 
 
 def test_explicit_from_earthengine_token():
